[Feat] 리뷰자 섹션 정보 가져오기

+++ BJJ_iOS/BJJ_iOS/Source/Network/API/MenuDetail/Model/MenuDetailModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct MenuDetailReviewInfo: Hashable, Codable {
     let reviewID: Int
     let comment: String
-    let reviewRating: Double
+    let reviewRating: Int
     let reviewImage: String
     let reviewLikeCount: Int
     let reviewCreatedDate: String
