+++ BJJ_iOS/BJJ_iOS/Source/Presentation/Tab/Home/MenuDetail/Model/MenuItem.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct MenuDetailModel: Hashable {
     let reviewComment: String
-    let reviewRating: Double
+    let reviewRating: Int
     let reviewImage: String
     let reviewLikedCount: Int
     let reviewCreatedDate: String
